+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/multiplatform/targets/HubdleKotlinMultiplatformNativeExtension.kt
@@ -70,7 +70,7 @@ constructor(
                 val tvosX64Main: KotlinSourceSet? = sourceSets.findByName("tvosX64Main")
                 val tvosSimulatorArm64Main: KotlinSourceSet? =
                     sourceSets.findByName("tvosSimulatorArm64Main")
-                val wasm32Main: KotlinSourceSet? = sourceSets.findByName("wasm32Main")
+                val wasmMain: KotlinSourceSet? = sourceSets.findByName("wasmMain")
                 val watchosArm32Main: KotlinSourceSet? = sourceSets.findByName("watchosArm32Main")
                 val watchosArm64Main: KotlinSourceSet? = sourceSets.findByName("watchosArm64Main")
                 val watchosX64Main: KotlinSourceSet? = sourceSets.findByName("watchosX64Main")
@@ -97,7 +97,7 @@ constructor(
                 val tvosX64Test: KotlinSourceSet? = sourceSets.findByName("tvosX64Test")
                 val tvosSimulatorArm64Test: KotlinSourceSet? =
                     sourceSets.findByName("tvosSimulatorArm64Test")
-                val wasm32Test: KotlinSourceSet? = sourceSets.findByName("wasm32Test")
+                val wasmTest: KotlinSourceSet? = sourceSets.findByName("wasmTest")
                 val watchosArm32Test: KotlinSourceSet? = sourceSets.findByName("watchosArm32Test")
                 val watchosArm64Test: KotlinSourceSet? = sourceSets.findByName("watchosArm64Test")
                 val watchosX64Test: KotlinSourceSet? = sourceSets.findByName("watchosX64Test")
@@ -123,7 +123,7 @@ constructor(
                         tvosArm64Main,
                         tvosX64Main,
                         tvosSimulatorArm64Main,
-                        wasm32Main,
+                        wasmMain,
                         watchosArm32Main,
                         watchosArm64Main,
                         watchosX64Main,
@@ -149,7 +149,7 @@ constructor(
                         tvosArm64Test,
                         tvosX64Test,
                         tvosSimulatorArm64Test,
-                        wasm32Test,
+                        wasmTest,
                         watchosArm32Test,
                         watchosArm64Test,
                         watchosX64Test,